refactor: rename canceled to cancelled
