+++ Modules/Devices/src/Devices.Domain/DomainEvents/Outgoing/IdentityDeletionCancelledDomainEvent.cs
@@ -4,7 +4,7 @@ namespace Backbone.Modules.Devices.Domain.DomainEvents.Outgoing;
 
 public class IdentityDeletionCancelledDomainEvent : DomainEvent
 {
-    public IdentityDeletionCancelledDomainEvent(string identityAddress) : base($"{identityAddress}/IdentityDeletionCanceled", randomizeId: true)
+    public IdentityDeletionCancelledDomainEvent(string identityAddress) : base($"{identityAddress}/IdentityDeletionCancelled", randomizeId: true)
     {
         IdentityAddress = identityAddress;
     }
